split out terminal client from dht client
